Refactor graph to take conditions into account
A graph can now be updated from a requirement or a package and a StoredNode
can store a requirement or a package, so that we don't have to resolve twice
a requirement when a condition is found.

To prevent infinite loop, the "conditions" keyword must be removed from the
package once stored, but there is a bug preventing using this method at the
moment so it must be fixed on another branch before.