Update Package creation process
Ensure that package instance just needs a mapping so that edition methods
can be safely used. The creation process is ensured by a new
"wiz.package.create" function.

Update "wiz.package.extract" function to just pass the variant identifier
to the create function and let it handle mismatch errors.

Remove "wiz.package.generate_identifier" function at it causes confusion
as to why we use it for definition identifiers when we need to have the
version. Instead, add a "version_identifier" to the Definition class.