Update requirement combination process
We cannot use only parent from distance mapping to determine the requirement
of a node as it automatically select the shortest path and eliminate
requirement from other parents which could have an effect in the graph.

So we use all existing (and root) parent requirements from the node to
process a reliable combined requirement.