Change strategy for handling package extraction error when updating graph
Instead of creating a dummy package which encapsulate the error, we store
the error in the graph and reference it to its parent identifier (even
if the parent is 'root'). This will help us to optimize the resolution
process as node variant with error could then be called once and skipped
from all other graph combinations.

Errors are being stored in list in the order of discovery so that we
keep a logical Breadth First order.