Fix logic to determine whether conflicted nodes must be sorted
Conflicting nodes were previously sorted only if the distance
mapping changed. However, this is not a good enough metric as
distance mapping might have been computed before entering the
conflict resolution loop, or after updating the list of
conflicts (by adding a new package to the graph).

To solve this issue, we added a new boolean value which default to
True (to always sort the conflicted nodes during the first loop),
and gets updated if a new node is added to the graph. This value is
added as an attribute to the Resolver class as is needs to keep
track of the value even if the current graph combination is
discarded.