Add versup to project for versioning
This adds [Versup](https://github.com/Svenito/versup) to the project to
replace the closed source usurp with an opensource alternative.

This commit adds a dependency to the `setup.py` file and an initial
config file replacing the current usurp config. The config will not
create a changelog and will replace the version numbers in
`release_nodes.rst` file as usurp did.

The existing tags work already with usurp, so the next version can be
set by running

    versup patch

or whatever increment it needs. To ensure you are going to get what you
expect, you can run

    versup patch -n

as a dry run. It will inform you of all changes that are going to be
made.

Closes #23