Update logic handling namespace counter.
Initial counter is passed by 'Resolver.compute_packages' as the Graph does
not have any way to distinguish implicit and explicit requests, so the
namespace extraction from initial requirements must be done upstream.

Within the graph, the counter is only updated when the node is created to
prevent nodes with unfulfilled conditions to influence the package
extraction process in any way.