Refactor version comparision
The comparison of versions when types match was a little hard to parse
and contained nested if statements and ternary if statements. This patch
refactors this into a single line that makes use of Python's ability to
handle booleans like ints to return the relavant result based on the
value comparison of the versions.